requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
docopt==0.6.2
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
jsonpickle==0.9.6
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
munch==2.3.2
numpy==1.14.5
pandas==0.24.2
protobuf==3.7.1
py-cpuinfo==5.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
sacred==0.7.4
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.3
Werkzeug==0.15.2
wrapt==1.11.1
